Change Turrent Command Name Not Requirement
