--- src/main/java/frc/robot/commands/turret/TurretFineAdjustment.java
@@ -1,60 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2020 Team 501 - The PowerKnights. All Rights Reserved.       */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the 2020 Team 501 - The PowerKnights BSD license    */
-/* file in the root directory of the project.                                 */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.commands.turret;
-
-import org.slf4j.Logger;
-
-import frc.robot.OI;
-import frc.robot.commands.PKCommandBase;
-import frc.robot.subsystems.turret.ITurretSubsystem;
-import frc.robot.subsystems.turret.TurretFactory;
-import riolog.RioLogger;
-
-public class TurretFineAdjustment extends PKCommandBase {
-
-  /** Our classes' logger **/
-  private static final Logger logger = RioLogger.getLogger(TurretFineAdjustment.class.getName());
-
-  private final OI oi;
-  private final ITurretSubsystem turret;
-
-  private double firstVal, lastVal;
-  private boolean passedInitialDeadband;
-
-  public TurretFineAdjustment() {
-    logger.info("constructing {}", getName());
-
-    oi = OI.getInstance();
-    turret = TurretFactory.getInstance();
-
-    logger.info("constructed");
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    firstVal = oi.getTurretFineAdjustment();
-    lastVal = firstVal;
-    passedInitialDeadband = false;
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-    double currentVal = oi.getTurretFineAdjustment();
-
-    if (!(passedInitialDeadband) && (Math.abs(currentVal - firstVal) / firstVal) <= 0.05) {
-      passedInitialDeadband = true;
-    }
-
-    if (passedInitialDeadband && (lastVal != currentVal)) {
-      turret.fineTuneTargetAngle(currentVal);
-    }
-  }
-
-}
